# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/cloud": {
            "get": {
                "description": "list cloud config",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "List",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/cloud/subscribe": {
            "post": {
                "description": "subscribe cloud",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "Subscribe",
                "parameters": [
                    {
                        "description": "body of subscribe cloud",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.cloudSubscribeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/cloud/{cid}": {
            "get": {
                "description": "get cloud pod",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "GetHttp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "cloud config id",
                        "name": "cid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/app.InferenceDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "bad_request_body"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.InferenceDTO": {
            "type": "object",
            "properties": {
                "access_url": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "inference_id": {
                    "type": "string"
                }
            }
        },
...
        "controller.cloudSubscribeRequest": {
            "type": "object",
            "properties": {
                "cloud_id": {
                    "type": "string"
                }
            }
        },
...
    }
}